+++ src/modules/tvShows.js
@@ -1,3 +1,4 @@
+import { popupArrange, test } from '../commentUtils.js';
 import { getTVShows } from './createTVShow.js';
 import { displayTVShows } from './displayTVShows.js';
 import { getLikes } from './getLikes.js';
@@ -8,6 +9,8 @@ export const show = async () => {
   const like = await getLikes();
   await displayTVShows(tvShow, like);
   likeCount();
+  test();
+  popupArrange();
 };
 
 export default show;
